Reorganize cell-range util functionality
